# web/openapi.yaml
components:
  schemas:
...
    APIKeyAccessWriteRequest:
      additionalProperties: false
      properties:
        access:
          items:
            $ref: "#/components/schemas/PermissionGrant"
          type: array
        admin:
          type: boolean
      required:
        - access
        - admin
      type: object
    APIKeyCreateRequest:
      additionalProperties: false
      properties:
        expires_at:
          format: date-time
          type:
            - string
            - "null"
        name:
          type: string
      required:
        - name
      type: object
    APIKeyListResponse:
      additionalProperties: false
      properties:
        rows:
          items:
            $ref: "#/components/schemas/APIKey"
          type: array
        total:
          format: int32
          type: integer
      required:
        - rows
        - total
      type: object
...
    AssetType:
      enum:
        - asset
        - photo
      type: string
...
    CreateAPIKeyData:
      additionalProperties: false
      properties:
        access:
          items:
            $ref: "#/components/schemas/PermissionGrant"
          type: array
        admin:
          type: boolean
        created_at:
          format: date-time
          type: string
        expires_at:
          format: date-time
          type:
            - string
            - "null"
        id:
          format: uuid
          type: string
        key_prefix:
          type: string
        last_used_at:
          format: date-time
          type:
            - string
            - "null"
        name:
          type: string
        secret:
          type: string
      required:
        - access
        - admin
        - created_at
        - id
        - key_prefix
        - name
        - secret
      type: object
...
    Problem:
      additionalProperties: false
      properties:
        code:
          type: string
        detail:
          type: string
        field_errors:
          items:
            $ref: "#/components/schemas/FieldError"
          type: array
        status:
          format: int32
          type: integer
        title:
          type: string
        type:
          type: string
      required:
        - code
        - detail
        - status
        - title
        - type
      type: object
...
  securitySchemes:
    apiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey
    sessionAuth:
      in: cookie
      name: woodgate_session
      type: apiKey
info:
  title: WoodGate API
  version: 0.1.0
openapi: 3.1.0
paths:
  /api-keys:
    get:
      operationId: listAPIKeys
      parameters:
        - explode: false
          in: query
          name: limit
          schema:
            format: int32
            minimum: 1
            type: integer
        - explode: false
          in: query
          name: offset
          schema:
            format: int32
            minimum: 0
            type: integer
        - explode: false
          in: query
          name: search
          schema:
            type: string
        - explode: false
          in: query
          name: sort
          schema:
            type: string
        - explode: false
          in: query
          name: order
          schema:
            enum:
              - asc
              - desc
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyListResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
    post:
      operationId: createAPIKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/APIKeyCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAPIKeyData"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
  /api-keys/{id}:
    delete:
      operationId: deleteAPIKey
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
    get:
      operationId: getAPIKey
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKey"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
    patch:
      operationId: patchAPIKey
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/APIKeyAccessWriteRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKey"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
  /assets:
    get:
      operationId: listAssets
      parameters:
        - explode: false
          in: query
          name: limit
          schema:
            format: int32
            minimum: 1
            type: integer
        - explode: false
          in: query
          name: offset
          schema:
            format: int32
            minimum: 0
            type: integer
        - explode: false
          in: query
          name: search
          schema:
            type: string
        - explode: false
          in: query
          name: sort
          schema:
            type: string
        - explode: false
          in: query
          name: order
          schema:
            enum:
              - asc
              - desc
            type: string
        - explode: false
          in: query
          name: type
          schema:
            $ref: "#/components/schemas/AssetType"
            enum:
              - asset
              - photo
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetListResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
    post:
      operationId: createAsset
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AssetCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
  /assets/{id}:
    delete:
      operationId: deleteAsset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
    get:
      operationId: getAsset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
    patch:
      operationId: patchAsset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AssetUpdateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
...
  /checkins/departments:
    get:
      operationId: listCheckinDepartments
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentOptionListResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
...
security:
  - apiKeyAuth: []
  - sessionAuth: []
servers:
  - url: /api/v1